npm run seed now works for organizations. And when organizations log in it dispalys organization.handlebar page
