+++ controllers/api/organization-routes.js
@@ -1,5 +1,4 @@
 const router = require("express").Router();
-const { User } = require("../../models");
 const Organization = require("../../models/organization");
 // const withAuth = require('../../utils/auth');
 
@@ -63,26 +62,6 @@ router.get("/signup", async (req, res) => {
   } catch (err) {
     console.log(err);
     res.status(404).json(err);
-  }
-});
-
-// Use withAuth middleware to prevent access to route
-router.get("/profile", async (req, res) => {
-  try {
-    // Find the logged in user based on the session ID
-    const orgData = await Organization.findByPk(req.session.id, {
-      attributes: { exclude: ["password"] },
-    });
-
-    const organization = orgData.get({ plain: true });
-
-    res.render("organization", {
-      ...organization,
-      logged_in: true,
-    });
-  } catch (err) {
-    console.log(err);
-    res.status(500).json(err);
   }
 });
