+++ controllers/home-routes.js
@@ -42,5 +42,26 @@ router.post("/logout", (req, res) => {
   }
 });
 
+// Use withAuth middleware to prevent access to route
+router.get("/profile", async (req, res) => {
+  try {
+    console.log(req.session.user_id)
+    // Find the logged in organization based on the session ID
+    const orgData = await Organization.findByPk(req.session.user_id, {
+      attributes: { exclude: ["password"] },
+    });
+
+    const organization = orgData.get({ plain: true });
+
+    res.render("organization", {
+      ...organization,
+      logged_in: true,
+    });
+  } catch (err) {
+    console.log(err);
+    res.status(500).json(err);
+  }
+});
+
 //add route for logout
 module.exports = router;
