+++ views/organization.handlebars
@@ -1,3 +1,3 @@
 {{! organization info page }}
-<h1>profile page</h1>
+<h2>profile page</h2>
 {{organization_name}}
