+++ views/organizationsignup.handlebars
@@ -30,7 +30,7 @@
       <input
         class="form-input"
         type="text"
-        id="volunteer-phone"
+        id="organization-phone"
         placeholder="Enter Organization Contact Phone Number"
       />
     </div>
