+++ views/volunteer-signup.handlebars
@@ -12,11 +12,11 @@
     </div>
     <div class="form-group">
       <label for="email-signup">Email</label>
-      <input class="form-input" type="text" id="volEmail" />
+      <input class="form-input" type="text" id="volemail" />
     </div>
     <div class="form-group">
-      <label for="password-signup">Phone Number</label>
-      <input class="form-input" type="password" id="volphone" />
+      <label for="phone-signup">Phone Number</label>
+      <input class="form-input" type="volunteer-phone" id="volphone" />
     </div>
     <div class="form-group">
       <label for="zipcode-signup">Zip code</label>
